backend/pyproject.toml
[tool.poetry]
name = "q-pay"
version = "0.1.0"
description = "Backend for Qpay p2p service"
authors = ["Sergey Sobolenko"]
license = "Proprietary"

packages = [
    {include = "apps"},
    {include = "config"},
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.10"
celery = "^5.3.6"
channels = "^4.0.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
openpyxl = "^3.1.2"
pandas = "^2.2.1"
httpx = "^0.27.0"
pyotp = "^2.9.0"
redis = "^5.0.1"
uvicorn = "^0.27.1"
channels-redis = "^4.2.0"
websockets = "^12.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
django-silk = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
